Now we can reboot the router from the Agent Console.
Once an agent initiates the 'reboot router' sequence, the button is disabled
until such time as the router comes back.

Once the router has been rebooted the agent will be notified on the screen.

The reboot router sequence will timeout after a specified time limit.
Currently it is set to 30 seconds until we have been notified of a better
timeout